requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
cachetools==4.2.1
    # via google-auth
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via typer
google-api-core==1.26.3
    # via google-api-python-client
google-api-python-client==2.1.0
    # via sheetfu
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth==1.28.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
googleapis-common-protos==1.53.0
    # via google-api-core
httplib2==0.19.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==2.10
    # via requests
oauth2client==4.1.3
    # via sheetfu
packaging==20.9
    # via google-api-core
protobuf==3.15.7
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
python-frontmatter==1.0.0
    # via -r requirements.in
python-slugify==4.0.1
    # via -r requirements.in
pytz==2021.1
    # via google-api-core
pyyaml==5.4.1
    # via python-frontmatter
requests==2.25.1
    # via google-api-core
rsa==4.7.2
    # via
    #   google-auth
    #   oauth2client
sheetfu==1.5.4
    # via -r requirements.in
six==1.15.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   oauth2client
    #   protobuf
text-unidecode==1.3
    # via python-slugify
typer==0.3.2
    # via -r requirements.in
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.4
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
